pyproject.toml
[project]
name = "shout-subgroup"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Richard Walker", email = "richard.walker.90@gmail.com" }
]
dependencies = [
    "python-telegram-bot==20.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/shout_subgroup"]
